store/test/test_views.py
from django.http import HttpRequest
from importlib import import_module
from django.conf import settings
from django.test import TestCase
from store.models import Category, product
from django.contrib.auth.models import User
from unittest import skip
from django.test import Client
from django.urls import reverse

from store.views import all_products


# @skip("demonstrating skipping")
# class TestSkip(TestCase):
#     def test_skip_example(self):
#         pass


class TestViewResponses(TestCase):
    def setUp(self):
        self.c = Client()
        Category.objects.create(name='django', slug='django')
        User.objects.create(username='admin')
        product.objects.create(category_id=1, title ='django beginners', created_by_id=1,
        slug='django-beginners', price='20.00', images='django')

    def test_url_allowed_hosts(self):
        response = self.c.get('/')
        self.assertEqual(response.status_code, 200)

    def test_product_detail_url(self):
        response = self.c.get(reverse('store:product_detail', args=['django-beginners']))
        self.assertEqual(response.status_code, 200)

    def test_category_detail_url(self):
        response = self.c.get(reverse('store:category_list', args=['django']))
        self.assertEqual(response.status_code, 200)


    def test_homepage_html(self):
        request = HttpRequest()
        engine = import_module(settings.SESSION_ENGINE)
        request.session = engine.SessionStore()
        response = all_products(request)
        html = response.content.decode('utf8')
        self.assertIn('<title>Home</title>', html)
        # print(html)

    def test_url_allowed_hosts(self):
        response = self.c.get('/', HTTP_HOST='noaddress.com')
        self.assertEqual(response.status_code, 400) 
    








